backend/requirements.txt
asgiref==3.11.1
coverage==7.13.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.8.0
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dotenv==1.2.1
sqlparse==0.5.5
